refactor(application): ui 패키지 분리
