+++ src/main/java/baseball/ui/ConsoleView.java
@@ -1,10 +1,8 @@
-package baseball.domain;
+package baseball.ui;
 
 import nextstep.utils.Console;
 
-import static baseball.domain.ConsoleMessage.*;
-
-public class ConsoleInOut {
+public class ConsoleView {
 
     public static final String ERROR_PREFIX = "[ERROR]";
 
@@ -24,11 +22,5 @@ public class ConsoleInOut {
 
     public static void printErrorMessage(String message) {
         System.out.printf("%s %s%n", ERROR_PREFIX, message);
-    }
-
-    public static void printErrorMessageByGameStatus(GameStatus gameStatus) {
-        if (gameStatus == GameStatus.ERROR) {
-            printErrorMessage(BASEBALL_GAME_RESTART_ERROR);
-        }
     }
 }
